Refactor: move cookie clearing logic into useRedirectHook
