+++ src/hooks/useRedirectHook.jsx
@@ -1,11 +1,20 @@
-import { useEffect, useState } from "react"
+import { useContext, useEffect, useState } from "react"
 import { useHistory } from "react-router-dom"
+import axios from 'axios'
+
+// Import context
+const { LoginContext } = require('../contexts/LoginContext')
+
+// constants
+const userIdKey = "userId"
+const BACKEND_URL = process.env.REACT_APP_BACKEND_URL
 
 const useRedirectHook = () => {
   const [shouldRedirect, setShouldRedirect] = useState(false)
   const [redirectUrl, setRedirectUrl] = useState('')
   const [redirectComponentState, setRedirectComponentState] = useState({})
   const history = useHistory()
+  const setLogoutState = useContext(LoginContext)
 
   useEffect(() => {
     if (shouldRedirect) {
@@ -28,10 +37,18 @@ const useRedirectHook = () => {
     setShouldRedirect(true)
   }
 
-  const setRedirectToLogout = () => {
-    setRedirectUrl("/")
-    setRedirectComponentState({ isFromSignOutButton: true })
-    setShouldRedirect(true)
+  const setRedirectToLogout = async () => {
+    try {
+      // Call the logout endpoint in the API server to clear the browser cookie
+      localStorage.removeItem(userIdKey)
+      await axios.get(`${BACKEND_URL}/auth/logout`)
+      setRedirectUrl("/")
+      setRedirectComponentState({ isFromSignOutButton: true })
+      setShouldRedirect(true)
+      setLogoutState()
+    } catch (err) {
+      console.error(err)
+    }
   }
 
   return { setRedirectToNotFound, setRedirectToPage, setRedirectToLogout }
